requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
autobahn==19.11.2
Automat==20.2.0
bcrypt @ file:///C:/ci/bcrypt_1597936263757/work
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1601563330930/work
django-allauth==0.44.0
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangocms-admin-style==2.0.2
enum34==1.1.10
Faker @ file:///tmp/build/80754af9/faker_1602623994968/work
Flask==1.1.2
gunicorn==20.0.4
heroku==0.1.4
heroku3==4.2.3
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiteconnect==3.9.2
MarkupSafe==1.1.1
numpy==1.20.1
oauthlib==3.1.0
p5py==1.0.0
pandas==1.2.3
pep517==0.10.0
Pillow==8.0.1
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32==300
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.2
validate-email==1.3
virtualenv==20.4.2
Werkzeug==1.0.1
whitenoise==5.2.0
wincertstore==0.2
zope.interface==5.2.0
